requirements.txt
fastapi
uvicorn[standard]
requests
pillow
opencv-python